refactor(frontend): split playtest box props
