+++ editor/crates/psxed-ui/src/lib.rs
@@ -33,16 +33,7 @@ mod animation_catalogue;
 use animation_catalogue::*;
 mod sector_inspector;
 use sector_inspector::*;
-mod workspace_dialogs;
-mod workspace_editing;
-mod workspace_geometry;
-mod workspace_input;
-mod workspace_painting;
-mod workspace_panels;
-mod workspace_selection;
-mod workspace_toolbars;
-mod workspace_ui_scene;
-mod workspace_viewport;
+mod workspace;
 pub use geometry::face_corner_world;
 use geometry::*;
 use ui_preview::*;
