+++ editor/crates/psxed-ui/src/workspace/mod.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+mod dialogs;
+mod editing;
+mod geometry;
+mod input;
+mod painting;
+mod panels;
+mod selection;
+mod toolbars;
+mod ui_scene;
+mod viewport;
