+++ editor/crates/psxed-ui/src/workspace/panels.rs
